+++ ampere/cache_plots.py
